common: substitute "off_t" with "long long int"
Modify libpmemlog and libpmemblk API to not use "off_t" type for
function arguments or returned value.

This modification is dictated by incompatible "off_t" definitions on
64-bit Linux and Windows operating systems.  On Windows, "off_t" is
always 32-bit long, which is not enough to hold offsets of large (>4GB)
pmem pool files.
To avoid potential problems with redefining "off_t" or with adding
additional typedefs, seems like the best choice is to use "long long"
type.

+++ src/examples/libpmemobj/manpage.c
@@ -44,7 +44,7 @@
 #include <libpmemobj.h>
 
 /* size of the pmemobj pool -- 1 GB */
-#define	POOL_SIZE ((off_t)(1 << 30))
+#define	POOL_SIZE ((size_t)(1 << 30))
 
 /* name of our layout in the pool */
 #define	LAYOUT_NAME "example_layout"
